requirements.txt
google-cloud-vision
Open-Tamil
python-docx
pillow
natsort
